Implement CSS selector matching and user-agent stylesheet defaults
- Added a new module for CSS selector matching in `selectors.rs`, including functions to handle various selector types and pseudo-classes.
- Introduced a user-agent stylesheet in `ua.rs` that defines default styles for HTML tags, including margins, font sizes, and display properties.
- Enhanced the selector matching logic to support complex selectors and pseudo-classes, ensuring accurate style application based on the DOM structure.

+++ kernel/src/boot/x86/userspace.rs
@@ -19,7 +19,7 @@ pub(super) fn start_userspace(_boot_info: &BootInfo, nogui: bool) -> ! {
 
     if !drivers::gpu::is_available() {
         drivers::gpu::bochs_vga::init(
-            framebuffer.addr as u32,
+            framebuffer.addr,
             framebuffer.width,
             framebuffer.height,
             framebuffer.pitch,
